Add files via upload
This will update the index.php file in all blog folders to match the template. Useful to run this if the template has changed and you don't want to update all of them manually. Basically added as a time saver because I'm lazy. I don't want to manually change all my files when I change the template.